feat: added enum descriptions for swagger. refactored filters out of program.cs
